requirements.txt
bitarray==2.8.1
bitstring==4.1.1
cffi==1.15.1
colorama==0.4.6
crayons==0.4.0
cryptography==41.0.3
numpy==1.25.2
opencv-python==4.8.0.76
piexif==1.1.3
Pillow==9.5.0
pycparser==2.21
pycryptodome==3.18.0
pydub==0.25.1
stegano==0.11.2
